Fix feeback manager question + feedback manager

+++ addons/shared_addons/modules/feedback_manager/controllers/feedback_manager.php
@@ -3,15 +3,15 @@
 defined('BASEPATH') or exit('No direct script access allowed');
 
 /**
- * Class Admin
- * @Description: This is class admin
+ * Class Feedback_manager
+ * @Description: This is class Feedback_manager
  * @Author: HoangThiTuanDung
- * @Company: framgia
  * @Date: 12/25/13
- * @Update:
+ * @Update: 12/25/13
  */
 class Feedback_manager extends Public_Controller {
 
+    // Validate
     protected $validation_rules = array(
         'title' => array(
             'field' => 'title',
@@ -49,11 +49,18 @@ class Feedback_manager extends Public_Controller {
             'rules' => ''
         ),
     );
-
+    
+    /**
+    * construct function
+    * load model and library
+    * check login
+    */
     public function __construct() {
+        
         parent::__construct();
-        if (!check_user_permission($this->current_user, $this->module, $this->permissions))
-            redirect();
+        
+        if (!check_user_permission($this->current_user, $this->module, $this->permissions)) redirect();
+        
         $this->template->set_layout('feedback_layout.html');
         $this->load->driver('cache', array('adapter' => 'apc', 'backup' => 'file'));
         $this->load->driver('Streams');
@@ -62,12 +69,22 @@ class Feedback_manager extends Public_Controller {
         $this->load->model(array('feedback_manager_m', 'feedbacktype/feedbacktype_m', 'feedbackquestion_m', 'feedback_manager_question/feedback_manager_question_m', 'feedbackuser/feedbackuser_m'));
         $this->lang->load('feedback_manager');
 
-        $feedback_manager_type = array('' => lang('feedback_manager:select_type'));
+        // Get feedback_manager_type list from cached to bidding select list
+        $feedback_manager_type = array(''  => lang('feedback_manager:select_type'));
         $feedback_manager_types = $this->streams->entries->get_entries(array('stream' => 'feedback_manager_type', 'namespace' => 'feedback_manager_types'));
         foreach ($feedback_manager_types['entries'] as $post) {
             $feedback_manager_type[$post['id']] = $post['title'];
         }
         $this->template->set('feedback_manager_types', $feedback_manager_type);
+
+
+        // Get apply list from cached to bidding select list
+        $apply = array(''  => lang('feedback_manager:select_apply'));
+        $users = $this->streams->entries->get_entries(array('stream' => 'profiles', 'namespace' => 'users'));
+        foreach ($users['entries'] as $post) {
+            $apply[$post['user_id']] = get_username_by_id($post['user_id']);
+        }
+        $this->template->set('users', $apply);  
     }
 
     /**
@@ -79,12 +96,13 @@ class Feedback_manager extends Public_Controller {
      * @Update: 11/21/13
      */
     public function index() {
+        //Search
         $where = "";
         if ($this->input->post('f_keywords')) {
             $where .= "`title` LIKE '%" . $this->input->post('f_keywords') . "%' ";
         }
 
-        // Get the latest team posts
+        // Get the latest feedback manager posts
         $posts = $this->streams->entries->get_entries(array(
             'stream' => 'feedback_manager',
             'namespace' => 'feedback_managers',
@@ -103,21 +121,22 @@ class Feedback_manager extends Public_Controller {
         $this->input->is_ajax_request() and $this->template->set_layout(false);
 
         $this->template
-                ->title($this->module_details['name'])
-                ->set_breadcrumb(lang('feedback_manager:feedback_manager_title'))
-                ->set('breadcrumb_title', $this->module_details['name'])
-                ->set_metadata('og:title', $this->module_details['name'], 'og')
-                ->set_metadata('og:type', 'feedback_manager', 'og')
-                ->set_metadata('og:url', current_url(), 'og')
-                ->set_metadata('og:description', $meta['description'], 'og')
-                ->set_metadata('description', $meta['description'])
-                ->set_metadata('keywords', $meta['keywords'])
-                ->append_js('module::feedback_manager_form.js')
-                ->append_js('module::statistics.js')
-                ->set_stream($this->stream->stream_slug, $this->stream->stream_namespace)
-                ->set('posts', $posts['entries'])
-                ->set('pagination', $posts['pagination']);
-        // ->set('feedback_manager_types', $this->cache->get('feedback_manager_types'));
+
+            ->title($this->module_details['name'])
+            ->set_breadcrumb(lang('feedback_manager:feedback_manager_title'))
+            ->set('breadcrumb_title', $this->module_details['name'])
+            ->set_metadata('og:title', $this->module_details['name'], 'og')
+            ->set_metadata('og:type', 'feedback_manager', 'og')
+            ->set_metadata('og:url', current_url(), 'og')
+            ->set_metadata('og:description', $meta['description'], 'og')
+            ->set_metadata('description', $meta['description'])
+            ->set_metadata('keywords', $meta['keywords'])
+            ->append_js('module::feedback_manager_form.js')
+            ->append_js('module::statistics.js')
+            ->set_stream($this->stream->stream_slug, $this->stream->stream_namespace)
+            ->set('posts', $posts['entries'])
+            ->set('pagination', $posts['pagination']);
+
         $this->input->is_ajax_request() ? $this->template->build('tables/posts') : $this->template->build('index');
     }
 
@@ -143,7 +162,7 @@ class Feedback_manager extends Public_Controller {
      * The delete function
      * @Description: This is delete function
      * @Parameter:
-     *      1. $id int The ID of the team post to delete
+     *      1. $id int The ID of the feedback manager post to delete
      * @Return: null
      * @Date: 11/21/13
      * @Update: 11/21/13
@@ -153,16 +172,21 @@ class Feedback_manager extends Public_Controller {
         if (!empty($ids)) {
             $post_names = array();
             $deleted_ids = array();
-            foreach ($ids as $id) {
-                if ($post = $this->feedback_manager_m->get($id)) {
-                    $postx = $this->feedbackuser_m->get_by(array('feedback_manager_id' => $id));
-                    $posty = $this->feedback_manager_question_m->get_by(array('feedback_manager_id' => $id));
-                    if ($postx == "" && $posty == "") {
-                        if ($this->feedback_manager_m->delete($id)) {
-                            $this->pyrocache->delete('feedback_manager_m');
-                            $post_names[] = $post->title;
-                            $deleted_ids[] = $id;
+
+            foreach ($ids as $id){
+                if ($post = $this->feedback_manager_m->get($id)){
+                    
+                    $postx = $this->feedbackuser_m->get_by(array('feedback_manager_id'=>$id));
+                    $posty = $this->feedback_manager_question_m->get_by(array('feedback_manager_id'=>$id));
+                    
+                    if ($this->feedback_manager_m->delete($id)){
+                        if($this->feedbackuser_m->delete($postx->id))
+                        {
+                            $this->pyrocache->delete('feedbackuser_m');
                         }
+                        $this->pyrocache->delete('feedback_manager_m');
+                        $post_names[] = $post->title;
+                        $deleted_ids[] = $id;
                     }
                 }
             }
@@ -204,11 +228,8 @@ class Feedback_manager extends Public_Controller {
     }
 
     /**
-     * The get_team_by_id function
-     * @Description: This is edit function
-     * @Parameter:
-     *      1. $id int The ID of the team post to get
-     * @Return: null
+     * The get_feedback_manager_by_id function
+     * @Parameter: $id
      * @Date: 11/21/13
      * @Update: 11/21/13
      */
@@ -223,10 +244,16 @@ class Feedback_manager extends Public_Controller {
         }
     }
 
-    public function get_question_manager($id) {
-        if (!$this->input->is_ajax_request())
-            redirect('feedback_manager');
-        if ($id != null && $id != "") {
+    /**
+     * The get_question_manager function
+     * @Parameter: $id
+     * @Date: 11/21/13
+     * @Update: 11/21/13
+     */
+    public function get_question_manager($id)
+    {
+        if(!$this->input->is_ajax_request()) redirect('feedback_manager');
+        if($id != null && $id != ""){
             $item = $this->feedbackquestion_m->get_question_list_by_fid($id);
             if (count($item) > 0)
                 echo json_encode($item);
@@ -254,27 +281,49 @@ class Feedback_manager extends Public_Controller {
         $this->form_validation->set_rules($rules);
 
         if ($this->form_validation->run()) {
-            $extra = array(
-                'title' => $this->input->post('title'),
-                'description' => $this->input->post('description'),
-                'start_date' => $this->input->post('start_date'),
-                'end_date' => $this->input->post('end_date'),
-                'type_id' => $this->input->post('type_id'),
-                'require' => $this->input->post('require'),
-                'status' => $this->input->post('status'),
-                'created' => date('Y-m-d H:i:s', now()),
-                'created_by' => $this->current_user->id
-            );
+            if(is_numeric($this->input->post('apply_id')))
+            {
+                $extra = array(
+                    'title' => $this->input->post('title'),
+                    'description' => $this->input->post('description'),
+                    'start_date' => $this->input->post('start_date'),
+                    'end_date' => $this->input->post('end_date'),
+                    'type_id' => $this->input->post('type_id'),
+                    'require' => $this->input->post('require'),
+                    'status' => $this->input->post('status'),
+                    'created' => date('Y-m-d H:i:s', now()),
+                    'created_by' => $this->current_user->id
+                );
 
+                if ($id = $this->streams->entries->insert_entry($_POST, 'feedback_manager', 'feedback_managers', array('created'), $extra)) {
+                    
+                    $extra2 = array(
+                        'user_id'               =>  $this->input->post('apply_id'),
+                        'feedback_manager_id'   =>  $id,
+                        'status'                => $this->input->post('status'),
+                        'date'                  => date('Y-m-d H:i:s', now()),
+                        'created'               => date('Y-m-d H:i:s', now()),
+                        'created_by'            => $this->current_user->id
+                    );
 
-            if ($id = $this->streams->entries->insert_entry($_POST, 'feedback_manager', 'feedback_managers', array('created'), $extra)) {
-                $this->pyrocache->delete_all('feedback_manager_m');
-                $message['status'] = 'success';
-                $message['message'] = str_replace("%s", $this->input->post('title'), lang('feedback_manager:post_add_success'));
-                Events::trigger('feedback_manager_created', $id);
-            } else {
+                    if ($id2 = $this->streams->entries->insert_entry($_POST, 'feedback_manager_user', 'feedback_manager_users', array('created'), $extra2)) {
+                        $this->pyrocache->delete_all('feedbackuser_m');
+                        Events::trigger('feedbackuser_add_created', $id2);
+                    }
+
+                    $this->pyrocache->delete_all('feedback_manager_m');
+                    $message['status'] = 'success';
+                    $message['message'] = str_replace("%s", $this->input->post('title'), lang('feedback_manager:post_add_success'));
+                    Events::trigger('feedback_manager_created', $id);
+                } else {
+                    $message['status'] = 'error';
+                    $message['message'] = lang('feedback_manager:post_add_error');
+                }
+            }
+            else
+            {
                 $message['status'] = 'error';
-                $message['message'] = lang('feedback_manager:post_add_error');
+                $message['message'] = lang('feedback_manager:validate_user_error');
             }
         } else {
             $message['status'] = 'error';
@@ -287,7 +336,7 @@ class Feedback_manager extends Public_Controller {
      * The edit function
      * @Description: This is edit function
      * @Parameter:
-     *      1. $id int The ID of the team post to edit
+     *      1. $id int The ID of the feedback manager post to edit
      * @Return: null
      * @Date: 11/21/13
      * @Update: 11/21/13
@@ -303,27 +352,49 @@ class Feedback_manager extends Public_Controller {
         $this->form_validation->set_rules($rules);
 
         if ($this->form_validation->run()) {
-            $author_id = empty($post->created_by) ? $this->current_user->id : $post->created_by;
-            $extra = array(
-                'title' => $this->input->post('title'),
-                'description' => $this->input->post('description'),
-                'start_date' => $this->input->post('start_date'),
-                'end_date' => $this->input->post('end_date'),
-                'type_id' => $this->input->post('type_id'),
-                'require' => $this->input->post('require'),
-                'status' => $this->input->post('status'),
-                'updated' => date('Y-m-d H:i:s', now()),
-                'created_by' => $author_id
-            );
+            if(is_numeric($this->input->post('apply_id')))
+            {
+                $author_id = empty($post->created_by) ? $this->current_user->id : $post->created_by;
+                $extra = array(
+                    'title' => $this->input->post('title'),
+                    'description' => $this->input->post('description'),
+                    'start_date' => $this->input->post('start_date'),
+                    'end_date' => $this->input->post('end_date'),
+                    'type_id' => $this->input->post('type_id'),
+                    'require' => $this->input->post('require'),
+                    'status' => $this->input->post('status'),
+                    'updated' => date('Y-m-d H:i:s', now()),
+                    'created_by' => $author_id
+                );
 
-            if ($this->streams->entries->update_entry($id, $_POST, 'feedback_manager', 'feedback_managers', array('updated'), $extra)) {
-                $message['status'] = 'success';
-                $message['message'] = str_replace("%s", $this->input->post('title'), lang('feedback_manager:edit_success'));
-                Events::trigger('feedback_manager_updated', $id);
-            } else {
-                $message['status'] = 'error';
-                $message['message'] = lang('feedback_manager:edit_error');
+                if ($this->streams->entries->update_entry($id, $_POST, 'feedback_manager', 'feedback_managers', array('updated'), $extra)) {
+                    $extra2 = array(
+                        'user_id'               =>  $this->input->post('apply_id'),
+                        'feedback_manager_id'   =>  $id,
+                        'status'                => $this->input->post('status'),
+                        'date'                  => date('Y-m-d H:i:s', now()),
+                        'updated'               => date('Y-m-d H:i:s', now()),
+                        'created_by'            => $this->current_user->id
+                    );
+                    $id2 = $this->feedbackuser_m->get_by(array('feedback_manager_id'=>$id))->id;
+                    if ($this->streams->entries->update_entry($id2, $_POST, 'feedback_manager_user', 'feedback_manager_users', array('updated'), $extra2)) {
+                        Events::trigger('feedback_manager_updated', $id2);
+                    }
+
+                    $message['status'] = 'success';
+                    $message['message'] = str_replace("%s", $this->input->post('title'), lang('feedback_manager:edit_success'));
+                    Events::trigger('feedback_manager_updated', $id);
+                } else {
+                    $message['status'] = 'error';
+                    $message['message'] = lang('feedback_manager:edit_error');
+                }
             }
+            else
+            {
+                $message['status'] = 'error';
+                $message['message'] = lang('feedback_manager:validate_user_error');
+            }
+            
         } else {
             $message['status'] = 'error';
             $message['message'] = lang('feedback_manager:validate_error');
